+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Added support for query `PeerStats`
 - Added support for query `PeerVersion`
 - Added support for query `NodeInfo`.
+- Added support for query `GetTransactionStatusInBlock`
 
 ## 3.0.0
 - Stronger typing for various places in the API.